requirements.txt
torch
faiss-cpu  # Use 'faiss-gpu' if running on GPU
sentence-transformers
numpy
jsonlines
PyMuPDF
ollama